+++ src/Data/KindID/Internal.hs
@@ -116,12 +116,25 @@ genKindIDs n = fmap KindID <$> V7.genUUIDs n
 nil :: KindID ""
 nil = KindID V7.nil
 {-# INLINE nil #-}
+{-# DEPRECATED nil "Use 'nilKindID' instead." #-}
+
+-- | The nil 'KindID'.
+nilKindID :: KindID ""
+nilKindID = KindID V7.nil
+{-# INLINE nilKindID #-}
 
 -- | Obtain a 'KindID' from a prefix and a 'UUID'.
 decorate :: forall prefix. (ToPrefix prefix, ValidPrefix (PrefixSymbol prefix))
          => UUID -> KindID prefix
 decorate = KindID
 {-# INLINE decorate #-}
+{-# DEPRECATED decorate "Use 'decorateKindID' instead." #-}
+
+-- | Obtain a 'KindID' from a prefix and a 'UUID'.
+decorateKindID :: forall prefix
+                . (ToPrefix prefix, ValidPrefix (PrefixSymbol prefix))
+               => UUID -> KindID prefix
+decorateKindID = KindID
 
 -- | Convert a 'KindID' to a 'TypeID'.
 toTypeID :: forall prefix. (ToPrefix prefix, ValidPrefix (PrefixSymbol prefix))